Make autosuggest onSelect cancellable
